[MIG] website_sale_address_format: Migration to 17.0
