+++ website_sale_address_format/README.rst
@@ -78,7 +78,7 @@ Credits
 Authors
 -------
 
-* Quartile Limited
+* Quartile
 
 Contributors
 ------------
