+++ website_sale_address_format/controllers/main.py
@@ -1,4 +1,4 @@
-# Copyright 2020 Quartile Limited
+# Copyright 2020 Quartile (https://www.quartile.co)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from odoo import http
