+++ website_sale_address_format/models/res_country.py
@@ -1,4 +1,4 @@
-# Copyright 2022-2022 Quartile Limited
+# Copyright 2022-2022 Quartile (https://www.quartile.co)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 import re
