+++ website_sale_address_format/static/description/index.html
@@ -418,7 +418,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <div class="section" id="authors">
 <h3><a class="toc-backref" href="#toc-entry-4">Authors</a></h3>
 <ul class="simple">
-<li>Quartile Limited</li>
+<li>Quartile</li>
 </ul>
 </div>
 <div class="section" id="contributors">
